Extracted SmartEnterFeature
